python-engine/requirements.txt
# HTTP
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Timezone
pytz==2024.1

# Database (direct mongo access for logging)
pymongo==4.6.1

# Type hints support
typing-extensions==4.9.0